feat: create SharedValidator with comum schemas
