+++ src/api/validators/ProductValidator.ts
@@ -42,18 +42,6 @@ class ProductValidator {
       throw new ApiError(400, err.message);
     });
   }
-
-  async showOrDelete(req: Request) {
-    const schema = Yup.object({
-      id: Yup.string()
-        .matches(/^[0-9a-fA-F]{24}$/, 'id must be a ObjectId')
-        .required(),
-    });
-
-    return schema.validate(req.params).catch(err => {
-      throw new ApiError(400, err.message);
-    });
-  }
 }
 
 export default new ProductValidator();
