+++ src/api/validators/RoleValidator.ts
@@ -38,18 +38,6 @@ class RoleValidator {
       throw new ApiError(400, err.message);
     });
   }
-
-  async showOrDelete(req: Request) {
-    const schema = Yup.object({
-      id: Yup.string()
-        .matches(/^[0-9a-fA-F]{24}$/, 'id must be a ObjectId')
-        .required(),
-    });
-
-    return schema.validate(req.params).catch(err => {
-      throw new ApiError(400, err.message);
-    });
-  }
 }
 
 export default new RoleValidator();
